pyproject.toml
[project]
name = "simulations"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
    { name = "Antoine Lefebvre-Brossard", email = "antoinelb@protonmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.0",
    "numpy>=2.0.0",
    "polars>=1.30.0",
    "pydantic>=2.11.5",
    "ruamel-yaml>=0.18.12",
    "tqdm>=4.67.1",
    "typer>=0.16.0",
    "uvicorn>=0.34.3",
    "starlette>=0.47.0",
    "plotly>=6.1.2",
    "kaleido==0.2.1",
    "pandas>=2.3.0",
]

[project.scripts]
swash-gui = "src.cli:run_cli"
swg = "src.cli:run_cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
  "src",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.1",
    "pytest-timeout>=2.4.0",
    "ruff>=0.11.12",
    "ty>=0.0.1a8",
]

[tool.black]
line_length = 79
target-version = ['py313']

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::FutureWarning",
  "ignore::UserWarning",
]
timeout = 60

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["src"]
